requirements.txt
asyncio~=3.4.3
discord.py[voice]~=1.7.3